refactor solana core test into integration with LOOP plugin
