+++ core/cmd/chains_commands_test.go
@@ -7,14 +7,11 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	solcfg "github.com/smartcontractkit/chainlink-solana/pkg/solana/config"
-
 	"github.com/smartcontractkit/chainlink-common/pkg/sqlutil"
 	client2 "github.com/smartcontractkit/chainlink-evm/pkg/client"
 	"github.com/smartcontractkit/chainlink/v2/core/cmd"
 	"github.com/smartcontractkit/chainlink/v2/core/internal/cltest"
 	"github.com/smartcontractkit/chainlink/v2/core/internal/testutils"
-	"github.com/smartcontractkit/chainlink/v2/core/internal/testutils/solanatest"
 	"github.com/smartcontractkit/chainlink/v2/core/services/chainlink"
 )
 
@@ -37,24 +34,5 @@ func TestShell_IndexEVMChains(t *testing.T) {
 	require.Len(t, chains, 1)
 	c := chains[0]
 	assert.Equal(t, strconv.Itoa(client2.NullClientChainID), c.ID)
-	assertTableRenders(t, r)
-}
-
-func TestShell_IndexSolanaChains(t *testing.T) {
-	t.Parallel()
-
-	id := solanatest.RandomChainID()
-	cfg := solcfg.TOMLConfig{
-		ChainID: &id,
-		Enabled: ptr(true),
-	}
-	app := solanaStartNewApplication(t, &cfg)
-	client, r := app.NewShellAndRenderer()
-
-	require.NoError(t, cmd.NewChainClient(client, "solana").IndexChains(cltest.EmptyCLIContext()))
-	chains := *r.Renders[0].(*cmd.ChainPresenters)
-	require.Len(t, chains, 1)
-	c := chains[0]
-	assert.Equal(t, id, c.ID)
 	assertTableRenders(t, r)
 }
